requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
sqlite3
werkzeug
requests==2.31.0
secure-smtplib
python-dotenv==1.0.0
secure-filename
json5==0.9.14
python-dateutil==2.8.2
uuid
base64
random
os-sys
smtplib
email
time